Update openocd script to use default password.
This change causes the default password (for FPGA generation) to be used when no other password is specified while still maintaining the ability to specify a different password when required (presumably for RTL simulation debug sessions and future hard targets that include the password authentication module.  

The `riscv authdata_write` command is always executed (with the assumption that execution is a nop-op on targets without a password auth module.